Copy the rails new info to both With/Without sections
Since the With Docker section is much longer due to the content that has to be copied into files, it felt really weird to be reading about `rails new` and that it would take awhile to run and output a lot of info. I included this with slightly different wording for the With Docker section since it uses `.` as the rails app name which means the current directory.